[feat] 쿠키에서 토큰을 가져와 userId를 구하는 방식 -> 쿠키에서 userId를 바로 받아오는 방식으로 변경
