+++ app/api/auth/user-id/route.ts
@@ -1,0 +1,16 @@
+import { NextRequest, NextResponse } from "next/server";
+import { cookies } from "next/headers";
+
+export async function GET(request: NextRequest) {
+  const cookieStore = cookies();
+  const userId = (await cookieStore).get("userId")?.value || null;
+
+  if (!userId) {
+    return NextResponse.json(
+      { error: "로그인이 필요합니다." },
+      { status: 401 }
+    );
+  }
+
+  return NextResponse.json({ userId });
+}
